+++ jackson/src/main/java/com/threedr3am/bug/jackson/package-info.java
@@ -1,10 +1,10 @@
 /**
  * @author threedr3am
  *
- * CVE-2020-8840    com.threedr3am.bug.jackson.JndiConverterPoc
+ * CVE-2020-8840    com.threedr3am.bug.jackson.rce.JndiConverterPoc
  * CVE-2019-20330   com.threedr3am.bug.jackson.EhcacheJndi2
- * CVE-2019-14379   com.threedr3am.bug.jackson.EhcacheJndi
- * CVE-2019-12384   com.threedr3am.bug.jackson.H2Rce
+ * CVE-2019-14379   com.threedr3am.bug.jackson.rce.EhcacheJndi
+ * CVE-2019-12384   com.threedr3am.bug.jackson.rce.H2Rce
  * CVE-2019-12086   com.threedr3am.bug.jackson.MysqlFileRead
  *
  *
