+++ jackson/src/main/java/com/threedr3am/bug/jackson/rce/H2Rce.java
@@ -1,4 +1,4 @@
-package com.threedr3am.bug.jackson;
+package com.threedr3am.bug.jackson.rce;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
