+++ jackson/src/main/java/com/threedr3am/bug/jackson/ssrf/JREJeditorPaneSSRFPoc.java
@@ -1,11 +1,13 @@
-package com.threedr3am.bug.jackson;
+package com.threedr3am.bug.jackson.ssrf;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.threedr3am.bug.common.server.HTTPServer;
 import java.io.IOException;
 
 /**
- * jackson-databind <= 2.9.10.3 and <= 2.10.2 RCE，需要开启DefaultType（JRE内置依赖）
+ * jackson-databind <= 2.9.10.3 and <= 2.10.2 RCE，需要开启DefaultType
+ *
+ * （todo JRE自带依赖） (Discovered by threedr3am) 这个还是蛮好的gadget
  *
  * @author threedr3am
  */
